pyproject.toml
[project]
name = "ld-mcp"
version = "0.1.0"
description = "MCP server for accessing W3C Linked Data specifications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.2",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "html-to-markdown>=2.20.0",
    "pydantic-settings>=2.12.0",
    "rdflib>=7.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
ld-mcp = "main:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line length handled separately

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (run with pytest -m slow)",
]
